New upstream version 2.8.8

+++ src/MantissaExponentConverter.cc
@@ -3,11 +3,11 @@
 *                                                              *
 * (C) 2008 Dirk Zimoch (dirk.zimoch@psi.ch)                    *
 *                                                              *
-* This is a custom exponential format converter for            * 
+* This is a custom exponential format converter for            *
 * StreamDevice.                                                *
 * The number is represented as two signed integers, mantissa   *
 * and exponent, like in +00011-01                              *
-* Please refer to the HTML files in ../doc/ for a detailed     *
+* Please refer to the HTML files in ../docs/ for a detailed    *
 * documentation.                                               *
 *                                                              *
 * If you do any changes in this file, you are not allowed to   *
@@ -39,7 +39,7 @@
 class MantissaExponentConverter : public StreamFormatConverter
 {
     virtual int parse(const StreamFormat&, StreamBuffer&, const char*&, bool);
-    virtual int scanDouble(const StreamFormat&, const char*, double&);
+    virtual ssize_t scanDouble(const StreamFormat&, const char*, double&);
     virtual bool printDouble(const StreamFormat&, StreamBuffer&, double);
 };
 
@@ -50,13 +50,13 @@ parse(const StreamFormat&, StreamBuffer&,
     return double_format;
 }
 
-int MantissaExponentConverter::
+ssize_t MantissaExponentConverter::
 scanDouble(const StreamFormat& fmt, const char* input, double& value)
 {
     int mantissa;
     int exponent;
     int length = -1;
-    
+
     sscanf(input, "%d%d%n", &mantissa, &exponent, &length);
     if (fmt.flags & skip_flag) return length;
     if (length == -1) return -1;
@@ -70,15 +70,15 @@ printDouble(const StreamFormat& fmt, StreamBuffer& output, double value)
     // Have to divide value into mantissa and exponent
     // precision field is number of characters in mantissa
     // number of characters in exponent is at least 2
-    int spaces;
+    ssize_t spaces;
     StreamBuffer buf;
     int prec = fmt.prec;
-    
+
     if (prec < 1) prec = 6;
     buf.print("%.*e", prec-1, fabs(value)/pow(10.0, prec-1));
     buf.remove(1,1);
     buf.remove(buf.find('e'),1);
-    
+
     spaces = fmt.width-buf.length();
     if (fmt.flags & (space_flag|sign_flag) || value < 0.0) spaces--;
     if (spaces < 0) spaces = 0;
